BUG: use size_t for conversion to numpy array
std::accumulate was implicitly using unsigned int as a template
parameter. The integer is not explicitly a size_t to ensure the type
is large enough.

+++ Wrapping/sitkNumpyArrayConversion.cxx
@@ -320,7 +320,7 @@ sitk_SetImageFromArray( PyObject *SWIGUNUSEDPARM(self), PyObject *args )
     size.push_back( sitkImage->GetNumberOfComponentsPerPixel() );
     }
 
-  len = std::accumulate( size.begin(), size.end(), 1u, std::multiplies<size_t>() );
+  len = std::accumulate( size.begin(), size.end(), size_t(1), std::multiplies<size_t>() );
   len *= pixelSize;
 
   if ( buffer_len != len )
